api/requirements.txt
tiktoken
pandas
fastapi
openai
modal-client
matplotlib
plotly
langchain
nltk
matplotlib